"Proper trash cans for the locks"
"Due to an oversight in the code that randomizes the key locations for Lt. Surge's gym, the first trash can can have a lock if any other trash can has the other lock. The intended behavior here is have the first trash can have the lock only if the second or fourth trash cans have the other lock." Fixed